v1.2
Work in Progress

Have set up a Svelte Dashboard and Supabase Database.

Now it is time to plug the python prediction scripts into the models.

This method will make the programme a lot better as it saves local space, and is pulling from a single source of truth rather than individual .csv spreadsheets inside the application. It is also sustainable long term, as I can upload the Supabase.